Removing binary artifacts & local datasets
