+++ catboost/dotnet/CatBoostNetTests/DownloadHelpers.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Net;
+using System.Text;
+
+namespace CatBoostNetTests
+{
+    public class DownloadHelpers
+    {
+        public static void DownloadDataset(string url, string outFile)
+        {
+            using (var client = new WebClient())
+            {
+                client.DownloadFile(url, outFile);
+            }
+        }
+    }
+}
