+++ catboost/dotnet/dotnet.sln
@@ -4,10 +4,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 VisualStudioVersion = 15.0.28010.2048
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CatBoostNet", "CatBoostNet\CatBoostNet.csproj", "{50D772D4-8573-4DAA-AB47-0DDAA5C271F3}"
+	ProjectSection(ProjectDependencies) = postProject
+		{9B7B9F03-8C4E-48B6-86D7-1857EE2A4C06} = {9B7B9F03-8C4E-48B6-86D7-1857EE2A4C06}
+	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProofOfConcept", "ProofOfConcept\ProofOfConcept.csproj", "{9AB66ABC-6886-4F78-ACEF-E799912FF73C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CatBoostNetTests", "CatBoostNetTests\CatBoostNetTests.csproj", "{04D6755C-1E6D-46A4-B188-3646CD88BB72}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CatBoostNetTests", "CatBoostNetTests\CatBoostNetTests.csproj", "{04D6755C-1E6D-46A4-B188-3646CD88BB72}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "LibraryBuilder", "LibraryBuilder\LibraryBuilder.vcxproj", "{9B7B9F03-8C4E-48B6-86D7-1857EE2A4C06}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -31,18 +34,6 @@ Global
 		{50D772D4-8573-4DAA-AB47-0DDAA5C271F3}.Release|x64.Build.0 = Release|Any CPU
 		{50D772D4-8573-4DAA-AB47-0DDAA5C271F3}.Release|x86.ActiveCfg = Release|Any CPU
 		{50D772D4-8573-4DAA-AB47-0DDAA5C271F3}.Release|x86.Build.0 = Release|Any CPU
-		{9AB66ABC-6886-4F78-ACEF-E799912FF73C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9AB66ABC-6886-4F78-ACEF-E799912FF73C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9AB66ABC-6886-4F78-ACEF-E799912FF73C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{9AB66ABC-6886-4F78-ACEF-E799912FF73C}.Debug|x64.Build.0 = Debug|Any CPU
-		{9AB66ABC-6886-4F78-ACEF-E799912FF73C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{9AB66ABC-6886-4F78-ACEF-E799912FF73C}.Debug|x86.Build.0 = Debug|Any CPU
-		{9AB66ABC-6886-4F78-ACEF-E799912FF73C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9AB66ABC-6886-4F78-ACEF-E799912FF73C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9AB66ABC-6886-4F78-ACEF-E799912FF73C}.Release|x64.ActiveCfg = Release|Any CPU
-		{9AB66ABC-6886-4F78-ACEF-E799912FF73C}.Release|x64.Build.0 = Release|Any CPU
-		{9AB66ABC-6886-4F78-ACEF-E799912FF73C}.Release|x86.ActiveCfg = Release|Any CPU
-		{9AB66ABC-6886-4F78-ACEF-E799912FF73C}.Release|x86.Build.0 = Release|Any CPU
 		{04D6755C-1E6D-46A4-B188-3646CD88BB72}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{04D6755C-1E6D-46A4-B188-3646CD88BB72}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{04D6755C-1E6D-46A4-B188-3646CD88BB72}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -55,6 +46,16 @@ Global
 		{04D6755C-1E6D-46A4-B188-3646CD88BB72}.Release|x64.Build.0 = Release|Any CPU
 		{04D6755C-1E6D-46A4-B188-3646CD88BB72}.Release|x86.ActiveCfg = Release|Any CPU
 		{04D6755C-1E6D-46A4-B188-3646CD88BB72}.Release|x86.Build.0 = Release|Any CPU
+		{9B7B9F03-8C4E-48B6-86D7-1857EE2A4C06}.Debug|Any CPU.ActiveCfg = Debug|Win32
+		{9B7B9F03-8C4E-48B6-86D7-1857EE2A4C06}.Debug|x64.ActiveCfg = Debug|x64
+		{9B7B9F03-8C4E-48B6-86D7-1857EE2A4C06}.Debug|x64.Build.0 = Debug|x64
+		{9B7B9F03-8C4E-48B6-86D7-1857EE2A4C06}.Debug|x86.ActiveCfg = Debug|Win32
+		{9B7B9F03-8C4E-48B6-86D7-1857EE2A4C06}.Debug|x86.Build.0 = Debug|Win32
+		{9B7B9F03-8C4E-48B6-86D7-1857EE2A4C06}.Release|Any CPU.ActiveCfg = Release|Win32
+		{9B7B9F03-8C4E-48B6-86D7-1857EE2A4C06}.Release|x64.ActiveCfg = Release|x64
+		{9B7B9F03-8C4E-48B6-86D7-1857EE2A4C06}.Release|x64.Build.0 = Release|x64
+		{9B7B9F03-8C4E-48B6-86D7-1857EE2A4C06}.Release|x86.ActiveCfg = Release|Win32
+		{9B7B9F03-8C4E-48B6-86D7-1857EE2A4C06}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
